+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.1.5
+- Added a button to view progress over 100 days
+- rewards button coming soon
+
 # 1.1.0
 - added streak display
 - add new streak animation!
@@ -15,5 +19,3 @@ If you fail to collect 5 stars in a day, your streak resets back to 0.
 
 Basic UI to track your current streak.
 
-# WARNING
-This mod is not 100% tested so there is a chance it might have bugs or not work at all!